requirements.txt
colorama
python-nmap
netifaces
